Validate
Leaf area index is one of the most important parameters in determining plant yields. On the other hand, agricultural scientists calculate this index using special hardware. But because of the global reach of the index in this method, the basis of validation is the data of MODIS product: MCD15A3H(500 M, 4 DAY).
 Random data analysis has been selected from all over the world. Product data is obtained from the following address:
https://lpdaacsvc.cr.usgs.gov/appeears/explore
Pearson correlation and mean error methods were used to check the validity of the data. The average error is obtained with this method:
MAE=(ABS(MODIS - S2)/C); MAE:  mean absolute error; C: count image (Picture of the same day) in the year: Sep 2019-Sep2020.

exam point of world:
long	lat
-3.787585	40.089773
-4.991665	33.647475
75.28038	24.862526
24.106951	-31.276369
-82.253509	34.582171
-101.92016	22.943836
-73.503856	3.385921
149.130662	-30.375819
22.50151	45.943571
11.464104	43.512098
11.664784	51.26062
1.229418	        44.669143
-122.048149	44.590588
-67.949633	-5.25761